fix: migrate back to the page routing in nextjs
ServerComponent from React are cool, but they require being well
supported by other dependencies, like UI components.

Currently, we don't need to scale a lot, so we'll keep our architecture
compatible with page routing. If there's a need to switch to a new component
pattern in the future, we'll take into consideration using
Next.js with the app routing framework.

Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>